firmware: jtag_scan()
- Separate implementation from vendor request handler.
- Remove redundant board-specific functions.
